ACMS-3623: PR feedback work.
