+++ modules/acquia_cms_common/acquia_cms_common.info.yml
@@ -5,7 +5,6 @@ type: module
 core_version_requirement: ~10.2.2
 dependencies:
   - acquia_purge:acquia_purge
-  - acquia_connector:acquia_connector
   - config_ignore:config_ignore
   - config_rewrite:config_rewrite
   - diff:diff
