+++ modules/acquia_cms_common/acquia_cms_common.services.yml
@@ -45,13 +45,3 @@ services:
       - '@theme_handler'
       - '@string_translation'
       - '@extension.list.module'
-
-  acquia_cms_common.telemetry:
-    class: Drupal\acquia_cms_common\EventSubscriber\KernelTerminate\AcquiaCmsTelemetry
-    arguments:
-      - '@acquia_connector.telemetry_service'
-      - '@config.factory'
-      - '@state'
-      - '%site.path%'
-    tags:
-      - { name: event_subscriber }
