+++ modules/acquia_cms_search/tests/src/Kernel/AcquiaFacetFacadeTest.php
@@ -24,8 +24,6 @@ class AcquiaFacetFacadeTest extends KernelTestBase {
     'search_api',
     'acquia_cms_search',
     'acquia_cms_common',
-    'acquia_connector',
-    'system',
   ];
 
   /**
